+++ Modules/Transaction/app/Http/Controllers/TransactionController.php
@@ -1,13 +1,14 @@
 <?php
 
-namespace App\Http\Controllers;
+namespace Modules\Transaction\Http\Controllers;
 
+use App\Http\Controllers\Controller;
 use Illuminate\Http\RedirectResponse;
 use Illuminate\Http\Request;
 use Inertia\Inertia;
 use Inertia\Response;
 
-class TransactionsController extends Controller
+class TransactionController extends Controller
 {
     public function index(): Response
     {
