+++ Modules/Transaction/app/Providers/RouteServiceProvider.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace Modules\Transactions\Providers;
+namespace Modules\Transaction\Providers;
 
 use Illuminate\Support\Facades\Route;
 use Illuminate\Foundation\Support\Providers\RouteServiceProvider as ServiceProvider;
@@ -34,7 +34,7 @@ class RouteServiceProvider extends ServiceProvider
      */
     protected function mapWebRoutes(): void
     {
-        Route::middleware('web')->group(module_path('Transactions', '/routes/web.php'));
+        Route::middleware('web')->group(module_path('Transaction', '/routes/web.php'));
     }
 
     /**
@@ -44,6 +44,6 @@ class RouteServiceProvider extends ServiceProvider
      */
     protected function mapApiRoutes(): void
     {
-        Route::middleware('api')->prefix('api')->name('api.')->group(module_path('Transactions', '/routes/api.php'));
+        Route::middleware('api')->prefix('api')->name('api.')->group(module_path('Transaction', '/routes/api.php'));
     }
 }
