+++ Modules/Transaction/composer.json
@@ -1,5 +1,5 @@
 {
-    "name": "nwidart/transactions",
+    "name": "nwidart/transaction",
     "description": "",
     "authors": [
         {
@@ -17,14 +17,14 @@
     },
     "autoload": {
         "psr-4": {
-            "Modules\\Transactions\\": "app/",
-            "Modules\\Transactions\\Database\\Factories\\": "database/factories/",
-            "Modules\\Transactions\\Database\\Seeders\\": "database/seeders/"
+            "Modules\\Transaction\\": "app/",
+            "Modules\\Transaction\\Database\\Factories\\": "database/factories/",
+            "Modules\\Transaction\\Database\\Seeders\\": "database/seeders/"
         }
     },
     "autoload-dev": {
         "psr-4": {
-            "Modules\\Transactions\\Tests\\": "tests/"
+            "Modules\\Transaction\\Tests\\": "tests/"
         }
     }
 }
