+++ Modules/Transaction/config/config.php
@@ -1,0 +1,5 @@
+<?php
+
+return [
+    'name' => 'Transaction',
+];
