--- Modules/Transactions/config/config.php
@@ -1,5 +1,0 @@
-<?php
-
-return [
-    'name' => 'Transactions',
-];
