--- Modules/Transactions/module.json
@@ -1,11 +1,0 @@
-{
-    "name": "Transactions",
-    "alias": "transactions",
-    "description": "",
-    "keywords": [],
-    "priority": 0,
-    "providers": [
-        "Modules\\Transactions\\Providers\\TransactionsServiceProvider"
-    ],
-    "files": []
-}
