+++ modules_statuses.json
@@ -1,4 +1,4 @@
 {
     "Initialize": true,
-    "Transactions": true
+    "Transaction": true
 }
